backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
google-genai==1.0.0
firebase-admin==6.6.0
sse-starlette==2.1.3
websockets==13.1
pydantic==2.9.0
python-multipart==0.0.12
google-auth==2.36.0
httpx==0.28.1
